Don't derive debug for error
Originally, ZipError derived debug so that it could be unwrapped.
Apparently, this is required for unwrapping a Result because, in the
even the unwrap panics, you have to have something to print to the
screen. I no longer unwrap ZipError for any reason, so I no longer need
the debug derivation.