Use serde
Apparently, my problem before was that the lat/long values were being
presented in my struct as floats and in the json as strings. This is
perfectly legal in Json.NET, but not here in Rust.

I'm also now using a rename instead of a string replace to get the
property names right.