Fix for lifetime error?
A client created inside the query function apparently blows up too soon
since the latest round of rustup-ing, either because of a change to
rustc (doubtful) or a change to the way hyper's client is build
(considered more likely).

The easy fix (and an ugly one, I think) was to just make the client
inside main and pass a reference to it to the query function instead.
Kinda hoping this is temporary, but not sure what other options I have.